+++ Solution/RiskOfChaos/Networking/Components/IsJumpingOnJumpPadTracker.cs
@@ -1,4 +1,5 @@
-﻿using RiskOfChaos.Utilities;
+﻿using RiskOfChaos.Patches;
+using RiskOfChaos.Utilities;
 using RoR2;
 using UnityEngine;
 using UnityEngine.Networking;
@@ -8,6 +9,22 @@ namespace RiskOfChaos.Networking.Components
     [RequireComponent(typeof(CharacterMotor))]
     public sealed class IsJumpingOnJumpPadTracker : NetworkBehaviour
     {
+        [SystemInitializer(typeof(BodyCatalog))]
+        static void Init()
+        {
+            foreach (GameObject bodyPrefab in BodyCatalog.allBodyPrefabs)
+            {
+                if (bodyPrefab.GetComponent<CharacterMotor>())
+                {
+                    bodyPrefab.AddComponent<IsJumpingOnJumpPadTracker>();
+
+#if DEBUG
+                    Log.Debug($"Added tracker to prefab: {bodyPrefab.name}");
+#endif
+                }
+            }
+        }
+
         [SyncVar]
         public bool IsJumping;
 
@@ -20,12 +37,29 @@ namespace RiskOfChaos.Networking.Components
 
         void OnEnable()
         {
+            JumpVolumeHooks.OnJumpVolumeJumpAuthority += onJumpVolumeJumpAuthority;
             _motor.onHitGroundAuthority += onHitGroundAuthority;
         }
 
         void OnDisable()
         {
+            JumpVolumeHooks.OnJumpVolumeJumpAuthority -= onJumpVolumeJumpAuthority;
             _motor.onHitGroundAuthority -= onHitGroundAuthority;
+        }
+
+        void onJumpVolumeJumpAuthority(JumpVolume jumpVolume, CharacterMotor jumpingCharacterMotor)
+        {
+            if (!jumpingCharacterMotor || jumpingCharacterMotor != _motor)
+                return;
+
+#if DEBUG
+            if (!IsJumping)
+            {
+                Log.Debug($"{Util.GetBestBodyName(gameObject)} started jumping on jump pad");
+            }
+#endif
+
+            CmdSetIsJumping(true);
         }
 
         void onHitGroundAuthority(ref CharacterMotor.HitGroundInfo hitGroundInfo)